requirements.txt
asgiref==3.8.1
chardet==5.2.0
Django==5.2.2
djangorestframework==3.16.0
numpy==2.3.0
pandas==2.3.0
pillow==11.2.1
psycopg2==2.9.10
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2025.2
reportlab==4.4.1
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
